Move /trending page inside dashboard layout
