+++ apps/dashboard/src/app/(dashboard)/trending/components/time-range-switcher.client.tsx
@@ -1,10 +1,7 @@
 "use client";
+import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import {
-  ToggleGroup,
-  ToggleGroupItem,
-} from "../../../@/components/ui/toggle-group";
-import type { TimeRange } from "../../../lib/search";
+import type { TimeRange } from "../../../../lib/search";
 
 export function TimeRangeSwitcher() {
   const router = useRouter();
