+++ apps/dashboard/src/app/(dashboard)/trending/loading.tsx
@@ -1,4 +1,4 @@
-import { Spinner } from "../../@/components/ui/Spinner/Spinner";
+import { Spinner } from "@/components/ui/Spinner/Spinner";
 
 export default function Loading() {
   return (
