+++ apps/dashboard/src/app/(dashboard)/trending/opengraph-image.png
